feat: refactor GitHub user router and add user retrieval endpoint
